types: add wlr_android_wlegl type
On Droidian and similar GNU/Linux distributions, we use the android_wlegl
wayland protocol to share wl_buffers around.

This is usually implemented directly in libhybris, and the global is created
during the egl_bind_display_call.

wlroots expects a wlr_buffer for pretty much everything nowadays, and if
we let libhybris handle that, we will end up with plain wl_buffers
sent by the client that we would need to wrap around and keep tracking (as
we don't control their lifecycle).

Instead, implement the server-side part of android_wlegl directly in
wlroots. The android_wlegl global is now created by us, and we can
create wlr_buffers directly when needed.

We also get to drop egl_bind_display() for free.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>